trie tableaue

+++ src/components/Projets/Projets.js
@@ -13,7 +13,12 @@ function Projets() {
     }
 
     const formatDate = (string) => {
-        var options = { year: 'numeric', month: 'long', day: 'numeric' };
+        var options = { 
+            year: 'numeric', 
+            month: 'numeric', 
+            day: 'numeric' 
+        };
+
         return new Date(string).toLocaleDateString([], options);
     }
